+++ common/buf/multi_buffer_test.go
@@ -33,7 +33,7 @@ func TestMultiBufferAppend(t *testing.T) {
 	b := New()
 	b.AppendBytes('a', 'b')
 	mb.Append(b)
-	assert(mb.Len(), Equals, 2)
+	assert(mb.Len(), Equals, int32(2))
 }
 
 func TestMultiBufferSliceBySizeLarge(t *testing.T) {
@@ -46,5 +46,5 @@ func TestMultiBufferSliceBySizeLarge(t *testing.T) {
 	mb.Append(lb)
 
 	mb2 := mb.SliceBySize(4 * 1024)
-	assert(mb2.Len(), Equals, 4*1024)
+	assert(mb2.Len(), Equals, int32(4*1024))
 }
